warcaby/backend/pyproject.toml
[tool.poetry]
name = "warcaby-be"
version = "0.1.0"
description = ""
authors = ["Mikuła <monikamikula77@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
pydantic = "^1.9.1"
django-picklefield = "^3.0.1"
pytest-django = "^4.5.2"
wheel = "^0.37.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
